// src/state/interfaces.tsx
export interface Shirt {
  id: number
  type: string
  color: string
  size: string
  brand: string
}
export interface Pants {
  id: number
  type: string
  color: string
  size: number
  brand: string
}

export interface Shoes {
  id: number
  type: string
  color: string
  size: number
  brand: string
}